requirements.txt
blinker==1.9.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
fonttools==4.60.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.6
numpy==2.3.3
packaging==25.0
pandas==2.3.3
pillow==11.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
six==1.17.0
tzdata==2025.2
urllib3==2.5.0
uv==0.8.24
websockets==15.0.1
Werkzeug==3.1.3
pydantic>=2.0.0
pydantic-settings>=2.0.0
rich>=13.0.0